+++ utils/microbatch_reddit_quiver_gloo.py
@@ -1,3 +1,9 @@
+'''
+sample : quiver 
+dataset: reddit
+getmicrobatch : yes
+'''
+
 import argparse
 import os